// cdm-frontend/src/interfaces/validateRuleInterface.ts
export interface ValidateRuleItem {
  vrfcSn: number;
  stdSeCd: string;
  levlSeq: number;
  vrfcTblNm: string;
  vrfcColNm: string;
  vrfcRulNm: string;
  rfrncNm: string;
  rfrncDtlNm: string;
  esntlYn: string;
  rfrncTblNm: string;
  fkNm: string;
  rfrncColNm: string;
  stdTrmId: string;
  rulAplcnNm: string;
  unitNm: string;
  scpSeqNm: string;
  prmCrtrNm: string;
  vrblCn: string;
  vrblDtlCn: string;
  vrblRsltCn: string;
  rgtrId: string;
  regDt: string;
  mdfrId: string;
  mdfcnDt: string;
}